# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A high-performance API for tracking user activity with advanced caching, rate limiting, and JWT authentication. Built with Go, Echo, PostgreSQL, and Redis.",
        "title": "User Activity Tracking System API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.nexmedis.com/support",
            "email": "support@nexmedis.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/usage/client/{client_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve daily API usage statistics for a specific client for the last 7 days",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Usage"
                ],
                "summary": "Get client usage statistics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Client ID",
                        "name": "client_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Client usage retrieved successfully",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/model.DailyUsage"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Client ID is required",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - JWT token required",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Client not found",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to get client usage",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/usage/daily": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve daily API usage statistics aggregated by client for the last 7 days. Results are cached for 1 hour.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Usage"
                ],
                "summary": "Get daily usage statistics",
                "responses": {
                    "200": {
                        "description": "Daily usage retrieved successfully",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/model.DailyUsage"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - JWT token required",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to get daily usage",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/usage/stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve overall API usage statistics including total requests in the last 24 hours, 7 days, and total number of clients",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Usage"
                ],
                "summary": "Get overall usage statistics",
                "responses": {
                    "200": {
                        "description": "Usage stats retrieved successfully",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "type": "object",
                                    "properties": {
                                        "timestamp": {
                                            "type": "string"
                                        },
                                        "total_clients": {
                                            "type": "integer"
                                        },
                                        "total_requests_24h": {
                                            "type": "integer"
                                        },
                                        "total_requests_7d": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "message": {
                                    "type": "string"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - JWT token required",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to get usage stats",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/usage/top": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve the top 3 clients with the highest number of API requests in the last 24 hours. Implements cache prefetching to ensure fresh data.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Usage"
                ],
                "summary": "Get top clients",
                "responses": {
                    "200": {
                        "description": "Top clients retrieved successfully",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/model.TopClient"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - JWT token required",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to get top clients",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "success": {
                                    "type": "boolean"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "model.DailyUsage": {
            "description": "Daily API usage statistics for a client",
            "type": "object",
            "properties": {
                "client_id": {
                    "description": "Client UUID",
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "client_name": {
                    "description": "Client name",
                    "type": "string",
                    "example": "John Doe"
                },
                "count": {
                    "description": "Number of API requests",
                    "type": "integer",
                    "example": 150
                },
                "date": {
                    "description": "Date (YYYY-MM-DD)",
                    "type": "string",
                    "example": "2025-01-15"
                }
            }
        },
...
        "model.TopClient": {
            "description": "Client with highest API usage",
            "type": "object",
            "properties": {
                "client_id": {
                    "description": "Client UUID",
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "client_name": {
                    "description": "Client name",
                    "type": "string",
                    "example": "John Doe"
                },
                "email": {
                    "description": "Client email",
                    "type": "string",
                    "example": "john.doe@example.com"
                },
                "total_requests": {
                    "description": "Total API requests",
                    "type": "integer",
                    "example": 500
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API Key for client authentication",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}